Built the FormControl for the Control Component. Using a service, the mat-form tells the service what to think, and that service is then called in the search response to update what columns of data are shown.
By default snippet, title and pageid are ALWAYS displayed for the table.

Not quite sure how to limit the number of rows that a mat-table will print ---

Likely lots of code and styling to update.

Still need to properly implement Error messages on SearchBar component, and remove the debugging erros and do fancy Errors on ControlComponent.